porting to .NET 5.0
