+++ Editor/equations/Common/UndoRedo/Text/TextAction.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace Editor
+﻿namespace Editor
 {
     public class TextAction : EquationAction
     {
@@ -17,6 +12,16 @@ namespace Editor
         public TextAction(ISupportsUndo executor)
             : base(executor)
         {            
+        }
+
+        public TextAction(ISupportsUndo executor, int index, string text, int[] formats, EditorMode[] modes, CharacterDecorationInfo[] decorations)
+            : base(executor)
+        {
+            Index = index;
+            Text = text;
+            Formats = formats;
+            Modes = modes;
+            Decorations = decorations;
         }
     }
 }
