+++ Editor/equations/Common/UndoRedo/Text/TextFormatAction.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace Editor
+﻿namespace Editor
 {
     public class TextFormatAction : EquationAction
     {
