+++ MathiVersity.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.30723.0
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31321.278
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Editor", "Editor\Editor.csproj", "{B19FF7B5-8771-4222-A539-496B58ECCC12}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Editor", "Editor\Editor.csproj", "{B19FF7B5-8771-4222-A539-496B58ECCC12}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -16,8 +16,8 @@ Global
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{B19FF7B5-8771-4222-A539-496B58ECCC12}.Debug|Any CPU.ActiveCfg = Debug|x86
-		{B19FF7B5-8771-4222-A539-496B58ECCC12}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
-		{B19FF7B5-8771-4222-A539-496B58ECCC12}.Debug|Mixed Platforms.Build.0 = Debug|x86
+		{B19FF7B5-8771-4222-A539-496B58ECCC12}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{B19FF7B5-8771-4222-A539-496B58ECCC12}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
 		{B19FF7B5-8771-4222-A539-496B58ECCC12}.Debug|x86.ActiveCfg = Debug|x86
 		{B19FF7B5-8771-4222-A539-496B58ECCC12}.Debug|x86.Build.0 = Debug|x86
 		{B19FF7B5-8771-4222-A539-496B58ECCC12}.Release|Any CPU.ActiveCfg = Release|x86
@@ -28,5 +28,8 @@ Global
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {8D728F5A-F3BB-434F-8A9F-5B2E4B13C740}
 	EndGlobalSection
 EndGlobal
